OmniVision-AI-main/requirements.txt
django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-environ==0.12.0
drf-spectacular==0.28.0
celery==5.5.3
redis==6.4.0
channels==4.3.1
channels-redis==4.3.0
whitenoise==6.11.0
django-cors-headers==4.9.0
django-filter==25.1
psycopg[binary]==3.2.12
opencv-python==4.11.0.86
pillow==11.2.1

